Fixed security hole in assessment Started subtab implementation for question browser in tests
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@11951 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ assessment/displaytempimage.php
@@ -10,11 +10,15 @@
 	{
 		$type = "jpeg";
 	}
-  header("Content-Type: image/$type");
-  header('Content-Length: '.filesize($image));
-  readfile($image);
-	if (is_file($image))
+	$size = getimagesize($image);
+	if (is_array($size) && (strpos($size["mime"], "image") !== FALSE))
 	{
-		unlink ($image);
+		header("Content-Type: image/$type");
+		header('Content-Length: '.filesize($image));
+		readfile($image);
+		if (is_file($image))
+		{
+			unlink ($image);
+		}
 	}
 ?>
